Renaming fixture files for clarity
