+++ end-to-end-tests/tests/regressions/formFlicker.spec.ts
@@ -16,7 +16,7 @@
  */
 
 import { getSidebarPage, runModViaQuickBar } from "../../utils";
-import { test, expect } from "../../fixtures/extensionBase";
+import { test, expect } from "../../fixtures/testBase";
 import { ActivateModPage } from "../../pageObjects/extensionConsole/modsPage";
 
 // @ts-expect-error -- https://youtrack.jetbrains.com/issue/AQUA-711/Provide-a-run-configuration-for-Playwright-tests-in-specs-with-fixture-imports-only
