+++ end-to-end-tests/tests/smoke/floatingActionButton.spec.ts
@@ -15,7 +15,7 @@
  * along with this program.  If not, see <http://www.gnu.org/licenses/>.
  */
 
-import { test, expect } from "../../fixtures/extensionBase";
+import { test, expect } from "../../fixtures/testBase";
 import { ActivateModPage } from "../../pageObjects/extensionConsole/modsPage";
 import { FloatingActionButton } from "../../pageObjects/floatingActionButton";
 import { getSidebarPage, isSidebarOpen } from "../../utils";
